requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
opencv-contrib-python==4.5.1.48
matplotlib==3.3.3
scikit-image==0.18.1
#mahotas==1.4.4
pandas==1.2.0
Pillow~=8.1.0